+++ libs/angular-components/community/components/cards/accordion/accordion-item-header/accordion-item-header.component.ts
@@ -7,13 +7,12 @@ import {
   ViewEncapsulation,
 } from "@angular/core";
 import { AccordionItemComponent } from "../accordion-item/accordion-item.component";
-import { ButtonComponent } from "../../../buttons/button/button.component";
 import { CardColors } from "../../card/card-colors.directive";
 import {
   CardHeaderComponent,
   CardHeaderVariant,
 } from "../../card/card-header/card-header.component";
-import { IconColor, IconComponent } from "@tehik-ee/tedi-angular/tedi";
+import { IconColor, IconComponent, ButtonComponent } from "@tehik-ee/tedi-angular/tedi";
 
 const WHITE_ICON_VARIANTS = ["brand", "brand-dark"];