+++ libs/angular-components/tedi/components/content/index.ts
@@ -1,1 +1,2 @@
 export * from "./list/list.component";
+export * from "./text-group/";