+++ libs/angular-components/tedi/components/content/text-group/text-group-value.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'tedi-text-group-value',
+  template: `<ng-content />`,
+  standalone: true,
+})
+export class TextGroupValueComponent {}