+++ libs/angular-components/tedi/components/form/number-field/number-field.component.ts
@@ -11,6 +11,7 @@ import {
   ViewChild,
 } from "@angular/core";
 import { ButtonComponent } from "../../buttons/button/button.component";
+import { TediTranslationPipe } from "../../../services/translation/translation.pipe";
 import { ComponentInputs } from "../../../types/inputs.type";
 import { IconComponent } from "../../base/icon/icon.component";
 import { TextComponent } from "../../base/text/text.component";
@@ -33,6 +34,7 @@ export type NumberFieldSize = "default" | "small";
     IconComponent,
     TextComponent,
     FeedbackTextComponent,
+    TediTranslationPipe
   ],
   providers: [
     {